+++ modules/client/client.module.spec.js
@@ -1,285 +1,223 @@
 const
 	assert = require('assert'),
-	co = require('co')
+	co = require('co'),
+	express = require('express'),
+	http = require('http'),
+	request = require('request-promise-native')
 
 module.exports = {
 	testMe: function() {
 		const instance = this
 		describe('client.module', function() {
-			let req = {
-					headers: {},
-					connection: {},
-					method: ''
-				},
-				res = {
-					headers: {},
-					getHeader: function(headerName) {
-					},
-					set: function(headerName, headerValue) {
-						res.fakeVar = true
-						if (!res.headers) {
-							res.headers = {}
-						}
-						res.headers[headerName] = headerValue
-					},
-					status: function(statusCode) {
-						res.fakeVar = true
-						if (typeof res.response === 'undefined') {
-							res.response = {}
-						}
-						res.response.statusCode = statusCode
-						return res
-					},
-					jsonTemplate: function(resolvePromise, jsonObject) {
-						res.fakeVar = true
-						if (typeof res.response === 'undefined') {
-							res.response = {}
-						}
-						res.response.jsonBody = jsonObject
-						if (typeof resolvePromise === 'function') {
-							resolvePromise()
-							return
-						}
-						return res
-					},
-					endTemplate: function(resolvePromise) {
-						if (typeof resolvePromise === 'function') {
-							resolvePromise()
-							return
-						}
-					},
-					redirectTemplate: function(resolvePromise, statusCode, route) {
-						res.fakeVar = true
-						if (typeof res.response === 'undefined') {
-							res.response = {}
-						}
-						res.response.statusCode = statusCode
-						res.response.redirectRoute = route
-						if (typeof resolvePromise === 'function') {
-							resolvePromise()
-							return
-						}
-						return res
-					}
-				},
-				next = function(resolvePromise, errorObject) {
-					next.fakeVar = true
-					if (errorObject) {
-						next.fail = true
-						next.errorStatus = errorObject.status
-						next.errorMessage = errorObject.message
-					} else {
-						next.fail = false
-					}
-					if (typeof resolvePromise === 'function') {
-						resolvePromise()
-					}
-				}
+			before(function() {
+				return co(function*() {
+					let app = express()
+					app.use((req, res, next) => {
+						req.isAuthenticated = () => req.query.isAuthenticated === 'true'
+						next()
+					})
+					app.use(instance.setDefaultsBeforeRequest.bind(instance)())
+					app.use([
+							'/',
+							'/testRoute',
+							'/authenticatedTestRoute',
+							'/unauthorizedRoute',
+							'/unauthorizedNonLayoutDirectRoute',
+							'/unauthorizedPageRedirectRoute'
+						],
+						(req, res) => res.status(200).end()
+					)
+					let server = http.createServer(app)
+					yield new Promise((resolve, reject) => {
+						server.listen(1110, () => {
+							resolve(true)
+						})
+					})
+					return true
+				})
+			})
 			it('should execute testSetDefaultsBeforeRequest successfully', function() {
-				instance.testSetDefaultsBeforeRequest(req, res, next)
+				instance.testSetDefaultsBeforeRequest()
 			})
 		})
 	},
-	testSetDefaultsBeforeRequest: function(req, res, next) {
-		const instance = this,
-			originalModuleConfig = JSON.parse(JSON.stringify(this.moduleConfig)),
-			setDefaultsBeforeRequest = this.setDefaultsBeforeRequest.bind(this)
+	testSetDefaultsBeforeRequest: function() {
+		const originalModuleConfig = JSON.parse(JSON.stringify(this.moduleConfig))
 		let {moduleConfig} = this,
-			changeableInstance = this
+			changeableInstance = this,
+			originalPaths = []
 		describe('client.module.setDefaultsBeforeRequest', function() {
+			before(function() {
+				originalPaths = changeableInstance.paths
+				moduleConfig.anonymousAccessRoutes = ['/', '/testRoute', '/unauthorizedPageRedirectRoute']
+				changeableInstance.paths = ['/', '/testRoute', '/authenticatedTestRoute', '/unauthorizedRoute', '/unauthorizedNonLayoutDirectRoute', '/unauthorizedPageRedirectRoute']
+				moduleConfig.nonLayoutDirectRoutes = ['/unauthorizedNonLayoutDirectRoute']
+				delete moduleConfig.notFoundRedirectRoutes
+				delete moduleConfig.layoutRoutes
+				delete moduleConfig.unauthorizedPageRedirectRoute
+				delete moduleConfig.redirectUnauthorizedPagesToNotFound
+			})
 			it('should execute successfully and return 404 if the route is not found and notFoundRedirectRoutes is not set', function() {
 				return co(function*() {
-					delete moduleConfig.notFoundRedirectRoutes
-					changeableInstance.paths = []
-					req.originalUrl = '/notFoundRoute'
-					yield (new Promise((resolve, reject) => {
-						res.end = res.endTemplate.bind(res, resolve)
-						setDefaultsBeforeRequest()(req, res, next.bind(next, resolve))
-					}))
-					moduleConfig.notFoundRedirectRoutes = originalModuleConfig.notFoundRedirectRoutes
+					let res = null
+					try {
+						yield request('http://127.0.0.1:1110/notFoundRoute', {resolveWithFullResponse: true})
+					} catch(e) {
+						res = e
+					}
 					assert.strictEqual(res.response.statusCode, 404, `bad value ${res.response.statusCode} res.response.statusCode, expected 404`)
 					return true
 				})
 			})
-			it('should execute successfully, return 302 and redirect to the correct route if the route is not found, only notFoundRedirectRoutes.default is set and the user is not authenticated', function() {
+			it('should execute successfully, return 200 and redirect to the correct route if the route is not found, only notFoundRedirectRoutes.default is set, the request is GET and the user is not authenticated', function() {
 				return co(function*() {
-					req.originalUrl = '/notFoundRoute'
-					req.isAuthenticated = () => false
 					moduleConfig.notFoundRedirectRoutes = {default: '/testRoute'}
-					yield (new Promise((resolve, reject) => {
-						res.redirect = res.redirectTemplate.bind(res, resolve)
-						setDefaultsBeforeRequest()(req, res, next.bind(next, resolve))
-					}))
-					delete req.isAuthenticated
-					moduleConfig.notFoundRedirectRoutes = originalModuleConfig.notFoundRedirectRoutes
-					assert.strictEqual(res.response.statusCode, 302, `bad value ${res.response.statusCode} res.response.statusCode, expected 302`)
-					assert.strictEqual(res.response.redirectRoute, '/testRoute', `bad value ${res.response.redirectRoute} for res.response.redirectRoute, expected '/testRoute'`)
+					let res = yield request('http://127.0.0.1:1110/notFoundRoute', {resolveWithFullResponse: true})
+					assert.strictEqual(res.statusCode, 200, `bad value ${res.statusCode} res.statusCode, expected 200`)
+					assert.strictEqual(res.request.href, 'http://127.0.0.1:1110/testRoute', `bad value ${res.request.href} for res.request.href, expected 'http://127.0.0.1:1110/testRoute'`)
 					return true
 				})
 			})
-			it('should execute successfully, return 302 and redirect to the correct route if the route is not found, only notFoundRedirectRoutes.default is set and the user is authenticated', function() {
+			it('should execute successfully, return 200 and redirect to the correct route if the route is not found, only notFoundRedirectRoutes.default is set, the request is GET and the user is authenticated', function() {
 				return co(function*() {
-					req.originalUrl = '/notFoundRoute'
-					req.isAuthenticated = () => true
-					moduleConfig.notFoundRedirectRoutes = {default: '/testRoute'}
-					yield (new Promise((resolve, reject) => {
-						res.redirect = res.redirectTemplate.bind(res, resolve)
-						setDefaultsBeforeRequest()(req, res, next.bind(next, resolve))
-					}))
-					delete req.isAuthenticated
-					moduleConfig.notFoundRedirectRoutes = originalModuleConfig.notFoundRedirectRoutes
-					assert.strictEqual(res.response.statusCode, 302, `bad value ${res.response.statusCode} res.response.statusCode, expected 302`)
-					assert.strictEqual(res.response.redirectRoute, '/testRoute', `bad value ${res.response.redirectRoute} for res.response.redirectRoute, expected '/testRoute'`)
+					let res = yield request('http://127.0.0.1:1110/notFoundRoute?isAuthenticated=true', {resolveWithFullResponse: true})
+					assert.strictEqual(res.statusCode, 200, `bad value ${res.statusCode} res.statusCode, expected 200`)
+					assert.strictEqual(res.request.href, 'http://127.0.0.1:1110/testRoute?isAuthenticated=true', `bad value ${res.request.href} for res.request.href, expected 'http://127.0.0.1:1110/testRoute?isAuthenticated=true'`)
 					return true
 				})
 			})
-			it('should execute successfully, return 302 and redirect to the correct route if the route is not found, notFoundRedirectRoutes.authenticated is set and the user is authenticated', function() {
+			it('should execute successfully and return 200 and redirect to the correct route if the route is not found, notFoundRedirectRoutes.authenticated is set, the request is get and the user is authenticated', function() {
 				return co(function*() {
-					req.originalUrl = '/notFoundRoute'
-					req.isAuthenticated = () => true
 					moduleConfig.notFoundRedirectRoutes = {authenticated: '/authenticatedTestRoute', default: '/testRoute'}
-					yield (new Promise((resolve, reject) => {
-						res.redirect = res.redirectTemplate.bind(res, resolve)
-						setDefaultsBeforeRequest()(req, res, next.bind(next, resolve))
-					}))
-					delete req.isAuthenticated
-					moduleConfig.notFoundRedirectRoutes = originalModuleConfig.notFoundRedirectRoutes
-					assert.strictEqual(res.response.statusCode, 302, `bad value ${res.response.statusCode} res.response.statusCode, expected 302`)
-					assert.strictEqual(res.response.redirectRoute, '/authenticatedTestRoute', `bad value ${res.response.redirectRoute} for res.response.redirectRoute, expected '/authenticatedTestRoute'`)
+					let res = yield request('http://127.0.0.1:1110/notFoundRoute?isAuthenticated=true', {resolveWithFullResponse: true})
+					assert.strictEqual(res.statusCode, 200, `bad value ${res.statusCode} res.statusCode, expected 200`)
+					assert.strictEqual(res.request.href, 'http://127.0.0.1:1110/authenticatedTestRoute?isAuthenticated=true', `bad value ${res.request.href} for res.request.href, expected 'http://127.0.0.1:1110/authenticatedTestRoute?isAuthenticated=true'`)
+					return true
+				})
+			})
+			it('should execute successfully and return 404 if the route is not found, the request is POST and the user is not authenticated', function() {
+				return co(function*() {
+					let res = null
+					try {
+						yield request('http://127.0.0.1:1110/notFoundRoute', {method: 'POST', resolveWithFullResponse: true})
+					} catch(e) {
+						res = e
+					}
+					assert.strictEqual(res.response.statusCode, 404, `bad value ${res.response.statusCode} res.response.statusCode, expected 404`)
+					return true
+				})
+			})
+			it('should execute successfully and return 404 if the route is not found, the request is POST and the user is authenticated', function() {
+				return co(function*() {
+					let res = null
+					try {
+						yield request('http://127.0.0.1:1110/notFoundRoute?isAuthenticated=true', {method: 'POST', resolveWithFullResponse: true})
+					} catch(e) {
+						res = e
+					}
+					assert.strictEqual(res.response.statusCode, 404, `bad value ${res.response.statusCode} res.response.statusCode, expected 404`)
 					return true
 				})
 			})
 			it('should execute successfully and return 401 if the user is not authenticated, the route is not anonymous, the route is not a layout one and the route is not a non-layout direct one', function() {
 				return co(function*() {
-					delete moduleConfig.anonymousAccessRoutes
-					delete moduleConfig.layoutRoutes
-					delete moduleConfig.nonLayoutDirectRoutes
-					req.originalUrl = '/unathorizedRoute'
-					changeableInstance.paths = ['/unathorizedRoute']
-					req.isAuthenticated = () => false
-					yield (new Promise((resolve, reject) => {
-						res.end = res.endTemplate.bind(res, resolve)
-						setDefaultsBeforeRequest()(req, res, next.bind(next, resolve))
-					}))
-					delete req.isAuthenticated
-					moduleConfig.anonymousAccessRoutes = originalModuleConfig.anonymousAccessRoutes
-					moduleConfig.layoutRoutes = originalModuleConfig.layoutRoutes
-					moduleConfig.nonLayoutDirectRoutes = originalModuleConfig.nonLayoutDirectRoutes
-					delete changeableInstance.paths
+					let res = null
+					try {
+						yield request('http://127.0.0.1:1110/unauthorizedRoute', {resolveWithFullResponse: true})
+					} catch(e) {
+						res = e
+					}
+					assert.strictEqual(res.response.statusCode, 401, `bad value ${res.response.statusCode} res.response.statusCode, expected 401`)
+					return true
+				})
+			})
+			it('should execute successfully and return 401 if the user is not authenticated, the route is not anonymous, the route is a non-layout direct one, unauthorizedPageRedirectRoute is not set and redirectUnauthorizedPagesToNotFound is not true', function() {
+				return co(function*() {
+					let res = null
+					try {
+						yield request('http://127.0.0.1:1110/unauthorizedNonLayoutDirectRoute', {resolveWithFullResponse: true})
+					} catch(e) {
+						res = e
+					}
 					assert.strictEqual(res.response.statusCode, 401, `bad value ${res.response.statusCode} res.response.statusCode, expected 401`)
 					return true
 				})
 			})
 			it('should execute successfully and return 401 if the user is not authenticated, the route is not anonymous, the route is a layout one, unauthorizedPageRedirectRoute is not set and redirectUnauthorizedPagesToNotFound is not true', function() {
 				return co(function*() {
-					delete moduleConfig.anonymousAccessRoutes
-					delete moduleConfig.layoutRoutes
-					delete moduleConfig.nonLayoutDirectRoutes
-					delete moduleConfig.unauthorizedPageRedirectRoute
-					delete moduleConfig.redirectUnauthorizedPagesToNotFound
-					req.originalUrl = '/unathorizedRoute'
-					req.method = 'GET'
-					changeableInstance.paths = ['/unathorizedRoute']
-					changeableInstance.layoutRoutes = ['/unathorizedRoute']
-					req.isAuthenticated = () => false
-					yield (new Promise((resolve, reject) => {
-						res.end = res.endTemplate.bind(res, resolve)
-						setDefaultsBeforeRequest()(req, res, next.bind(next, resolve))
-					}))
-					delete req.isAuthenticated
-					moduleConfig.anonymousAccessRoutes = originalModuleConfig.anonymousAccessRoutes
-					moduleConfig.layoutRoutes = originalModuleConfig.layoutRoutes
-					moduleConfig.nonLayoutDirectRoutes = originalModuleConfig.nonLayoutDirectRoutes
-					moduleConfig.unauthorizedPageRedirectRoute = originalModuleConfig.unauthorizedPageRedirectRoute
-					moduleConfig.redirectUnauthorizedPagesToNotFound = originalModuleConfig.redirectUnauthorizedPagesToNotFound
-					delete changeableInstance.paths
-					delete changeableInstance.layoutRoutes
+					changeableInstance.layoutRoutes = ['/unauthorizedRoute']
+					let res = null
+					try {
+						yield request('http://127.0.0.1:1110/unauthorizedRoute', {resolveWithFullResponse: true})
+					} catch(e) {
+						res = e
+					}
 					assert.strictEqual(res.response.statusCode, 401, `bad value ${res.response.statusCode} res.response.statusCode, expected 401`)
 					return true
 				})
 			})
-			it('should execute successfully, return 302 and redirect to the correct route if the user is not authenticated, the route is not anonymous, the route is a layout one and unauthorizedPageRedirectRoute is set', function() {
+			it('should execute successfully, return 200 and redirect to the correct route if the user is not authenticated, the route is not anonymous, the route is a layout one and unauthorizedPageRedirectRoute is set', function() {
 				return co(function*() {
-					delete moduleConfig.anonymousAccessRoutes
-					delete moduleConfig.layoutRoutes
-					delete moduleConfig.nonLayoutDirectRoutes
-					moduleConfig.unauthorizedPageRedirectRoute = '/unauthrozidPageRedirectRoute'
-					delete moduleConfig.redirectUnauthorizedPagesToNotFound
-					req.originalUrl = '/unathorizedRoute'
-					req.method = 'GET'
-					changeableInstance.paths = ['/unathorizedRoute']
-					changeableInstance.layoutRoutes = ['/unathorizedRoute']
-					req.isAuthenticated = () => false
-					yield (new Promise((resolve, reject) => {
-						res.redirect = res.redirectTemplate.bind(res, resolve)
-						setDefaultsBeforeRequest()(req, res, next.bind(next, resolve))
-					}))
-					delete req.isAuthenticated
-					moduleConfig.anonymousAccessRoutes = originalModuleConfig.anonymousAccessRoutes
-					moduleConfig.layoutRoutes = originalModuleConfig.layoutRoutes
-					moduleConfig.nonLayoutDirectRoutes = originalModuleConfig.nonLayoutDirectRoutes
-					moduleConfig.unauthorizedPageRedirectRoute = originalModuleConfig.unauthorizedPageRedirectRoute
-					moduleConfig.redirectUnauthorizedPagesToNotFound = originalModuleConfig.redirectUnauthorizedPagesToNotFound
-					delete changeableInstance.paths
-					delete changeableInstance.layoutRoutes
-					assert.strictEqual(res.response.statusCode, 302, `bad value ${res.response.statusCode} res.response.statusCode, expected 404`)
+					moduleConfig.unauthorizedPageRedirectRoute = '/unauthorizedPageRedirectRoute'
+					let res = yield request('http://127.0.0.1:1110/unauthorizedRoute', {resolveWithFullResponse: true})
+					assert.strictEqual(res.statusCode, 200, `bad value ${res.statusCode} res.statusCode, expected 200`)
 					assert.strictEqual(
-						res.response.redirectRoute,
-						'/unauthrozidPageRedirectRoute',
-						`bad value ${res.response.redirectRoute} for res.response.redirectRoute, expected '/unauthrozidPageRedirectRoute'`
+						res.request.href,
+						'http://127.0.0.1:1110/unauthorizedPageRedirectRoute',
+						`bad value ${res.request.href} for res.request.href, expected 'http://127.0.0.1:1110/unauthorizedPageRedirectRoute'`
 					)
 					return true
 				})
 			})
-			it('should execute successfully, return 302 and redirect to the correct route if the user is not authenticated, the route is not anonymous, the route is a layout one, unauthorizedPageRedirectRoute is not set and redirectUnauthorizedPagesToNotFound is true', function() {
+			it('should execute successfully, return 200 and redirect to the correct route if the user is not authenticated, the route is not anonymous, the route is a non-layout direct one and unauthorizedPageRedirectRoute is set', function() {
 				return co(function*() {
-					delete moduleConfig.anonymousAccessRoutes
-					delete moduleConfig.layoutRoutes
-					delete moduleConfig.nonLayoutDirectRoutes
+					let res = yield request('http://127.0.0.1:1110/unauthorizedNonLayoutDirectRoute', {resolveWithFullResponse: true})
+					assert.strictEqual(res.statusCode, 200, `bad value ${res.statusCode} res.statusCode, expected 200`)
+					assert.strictEqual(
+						res.request.href,
+						'http://127.0.0.1:1110/unauthorizedPageRedirectRoute',
+						`bad value ${res.request.href} for res.request.href, expected 'http://127.0.0.1:1110/unauthorizedPageRedirectRoute'`
+					)
+					return true
+				})
+			})
+			it('should execute successfully, return 200 and redirect to the correct route if the user is not authenticated, the route is not anonymous, the route is a layout one, unauthorizedPageRedirectRoute is not set and redirectUnauthorizedPagesToNotFound is true', function() {
+				return co(function*() {
 					delete moduleConfig.unauthorizedPageRedirectRoute
 					moduleConfig.redirectUnauthorizedPagesToNotFound = true
-					moduleConfig.notFoundRedirectRoutes = {default: '/testRoute'}
-					req.originalUrl = '/unathorizedRoute'
-					req.method = 'GET'
-					changeableInstance.paths = ['/unathorizedRoute']
-					changeableInstance.layoutRoutes = ['/unathorizedRoute']
-					req.isAuthenticated = () => false
-					yield (new Promise((resolve, reject) => {
-						res.redirect = res.redirectTemplate.bind(res, resolve)
-						setDefaultsBeforeRequest()(req, res, next.bind(next, resolve))
-					}))
-					delete req.isAuthenticated
-					moduleConfig.anonymousAccessRoutes = originalModuleConfig.anonymousAccessRoutes
-					moduleConfig.layoutRoutes = originalModuleConfig.layoutRoutes
-					moduleConfig.nonLayoutDirectRoutes = originalModuleConfig.nonLayoutDirectRoutes
-					moduleConfig.unauthorizedPageRedirectRoute = originalModuleConfig.unauthorizedPageRedirectRoute
-					moduleConfig.redirectUnauthorizedPagesToNotFound = originalModuleConfig.redirectUnauthorizedPagesToNotFound
-					moduleConfig.notFoundRedirectRoutes = originalModuleConfig.notFoundRedirectRoutes
-					delete changeableInstance.paths
-					delete changeableInstance.layoutRoutes
-					assert.strictEqual(res.response.statusCode, 302, `bad value ${res.response.statusCode} res.response.statusCode, expected 302`)
-					assert.strictEqual(res.response.redirectRoute, '/testRoute', `bad value ${res.response.redirectRoute} for res.response.redirectRoute, expected '/testRoute'`)
+					let res = yield request('http://127.0.0.1:1110/unauthorizedRoute', {resolveWithFullResponse: true})
+					assert.strictEqual(res.statusCode, 200, `bad value ${res.statusCode} res.statusCode, expected 200`)
+					assert.strictEqual(res.request.href, 'http://127.0.0.1:1110/testRoute', `bad value ${res.request.href} for res.request.href, expected 'http://127.0.0.1:1110/testRoute'`)
+					return true
+				})
+			})
+			it('should execute successfully, return 200 and redirect to the correct route if the user is not authenticated, the route is not anonymous, the route is a non-layout direct one, unauthorizedPageRedirectRoute is not set and redirectUnauthorizedPagesToNotFound is true', function() {
+				return co(function*() {
+					let res = yield request('http://127.0.0.1:1110/unauthorizedNonLayoutDirectRoute', {resolveWithFullResponse: true})
+					assert.strictEqual(res.statusCode, 200, `bad value ${res.statusCode} res.statusCode, expected 200`)
+					assert.strictEqual(
+						res.request.href,
+						'http://127.0.0.1:1110/testRoute',
+						`bad value ${res.request.href} for res.request.href, expected 'http://127.0.0.1:1110/testRoute'`
+					)
 					return true
 				})
 			})
 			it('should execute successfully and continue to next if the user is not authenticated, but the route is anonymous', function() {
 				return co(function*() {
-					moduleConfig.anonymousAccessRoutes = ['/unathorizedRoute']
-					changeableInstance.paths = ['/unathorizedRoute']
-					req.originalUrl = '/unathorizedRoute'
-					req.isAuthenticated = () => false
-					next.fail = null
-					yield (new Promise((resolve, reject) => {
-						setDefaultsBeforeRequest()(req, res, next.bind(next, resolve))
-					}))
-					delete req.isAuthenticated
-					moduleConfig.anonymousAccessRoutes = originalModuleConfig.anonymousAccessRoutes
-					delete changeableInstance.paths
-					assert.strictEqual(next.fail, false, `bad value ${next.fail} for next.fail, expected false`)
-					assert.strictEqual(req.locals.error, null, `bad value ${req.locals.error} for req.locals.error, expected null`)
-					assert.strictEqual(req.locals.originalUrl, '/unathorizedRoute', `bad value ${req.locals.originalUrl} for req.locals.originalUrl, expected '/unathorizedRoute'`)
+					let res = yield request('http://127.0.0.1:1110/', {resolveWithFullResponse: true})
+					assert.strictEqual(res.statusCode, 200, `bad value ${res.statusCode} res.statusCode, expected 200`)
 					return true
 				})
+			})
+			after(function() {
+				changeableInstance.paths = originalPaths
+				moduleConfig.anonymousAccessRoutes = originalModuleConfig.anonymousAccessRoutes
+				moduleConfig.notFoundRedirectRoutes = originalModuleConfig.notFoundRedirectRoutes
+				moduleConfig.layoutRoutes = originalModuleConfig.layoutRoutes
+				moduleConfig.nonLayoutDirectRoutes = originalModuleConfig.nonLayoutDirectRoutes
+				moduleConfig.unauthorizedPageRedirectRoute = originalModuleConfig.unauthorizedPageRedirectRoute
+				moduleConfig.redirectUnauthorizedPagesToNotFound = originalModuleConfig.redirectUnauthorizedPagesToNotFound
 			})
 		})
 	}