+++ modules/client/index.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const ClientModule = require('./client.module.js'),
-	BaseClientComponent = require('./base-client.component.js')
+const ClientModule = require('./client.module'),
+	BaseClientComponent = require('./base-client.component')
 
 module.exports = {ClientModule, BaseClientComponent}